+++ checkbox.v
@@ -94,7 +94,7 @@ fn (mut b CheckBox) propose_size(w, h int) (int,int) {
 }
 
 fn (mut b CheckBox) draw() {
-	b.ui.gg.draw_rect(b.x, b.y, check_mark_size, check_mark_size, gx.white) // progress_bar_color)
+	b.ui.gg.draw_rect(f32(b.x), f32(b.y), check_mark_size, check_mark_size, gx.white) // progress_bar_color)
 	// b.ui.gg.draw_empty_rect(b.x, b.y, check_mark_size, check_mark_size, cb_border_color)
 	draw_inner_border(false, b.ui.gg, b.x, b.y, check_mark_size, check_mark_size)
 	// Draw X (TODO draw a check mark instead)
@@ -109,7 +109,7 @@ fn (mut b CheckBox) draw() {
 		b.ui.gg.draw_line_c(x0, y1, x0+check_mark_size -4, y0, gx.black)
 		b.ui.gg.draw_line_c(0.5+x0, y1, -3.5+x0+check_mark_size, y0, gx.black)
 		*/
-		b.ui.gg.draw_image(b.x + 3, b.y + 3, 8, 8, b.ui.cb_image)
+		b.ui.gg.draw_image(f32(b.x) + 3, f32(b.y) + 3, 8, 8, b.ui.cb_image)
 	}
 	// Text
 	b.ui.ft.draw_text(b.x + check_mark_size + 5, b.y, b.text, btn_text_cfg)
