+++ rectangle.v
@@ -74,15 +74,15 @@ fn (mut r Rectangle) propose_size(w, h int) (int,int) {
 
 fn (mut r Rectangle) draw() {
 	if r.radius > 0 {
-		r.ui.gg.draw_rounded_rect(r.x, r.y, r.width, r.height, r.radius, r.color)
+		r.ui.gg.draw_rounded_rect(f32(r.x), f32(r.y), f32(r.width), f32(r.height), f32(r.radius), r.color)
 		if r.border {
-			r.ui.gg.draw_empty_rounded_rect(r.x, r.y, r.width, r.height, r.radius, r.border_color)
+			r.ui.gg.draw_empty_rounded_rect(f32(r.x), f32(r.y), f32(r.width), f32(r.height), f32(r.radius), r.border_color)
 		}
 	}
 	else {
-		r.ui.gg.draw_rect(r.x, r.y, r.width, r.height, r.color)
+		r.ui.gg.draw_rect(f32(r.x), f32(r.y), f32(r.width), f32(r.height), r.color)
 		if r.border {
-			r.ui.gg.draw_empty_rect(r.x, r.y, r.width, r.height, r.border_color)
+			r.ui.gg.draw_empty_rect(f32(r.x), f32(r.y), f32(r.width), f32(r.height), r.border_color)
 		}
 	}
 	text_cfg := gx.TextCfg{
